requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
Babel==2.3.4              # via babeldjango, django-prices
BabelDjango==0.2.2        # via django-prices
boto==2.43.0
braintree==3.30.0         # via django-payments
cffi==1.8.3               # via cryptography
cryptography==1.5.3       # via django-payments
cssselect==1.0.0          # via premailer
cssutils==1.0.1           # via premailer
dj-database-url==0.4.1
dj-email-url==0.0.10
django-absolute==0.3      # via django-emailit
django-bootstrap3==7.1.0
django-cache-url==1.3.0
django-compressor==2.1
django-countries==4.0
django-emailit==0.2.2
django-materializecss-form==1.0.1
django-model-utils==2.6
django-mptt==0.8.6
django-payments==0.9.5
django-prices-openexchangerates==0.1.12
django-prices==0.4.13
django-redis==4.6.0
django-selectable==0.9.0
django-storages==1.5.1
django-versatileimagefield==1.6
django-webpack-loader==0.4.1
Django==1.10.3            # via django-absolute, django-emailit, django-model-utils, django-payments, django-prices, django-prices-openexchangerates, jsonfield
djangorestframework==3.5.2
fabric3==1.12.post1
fake-factory==0.7.2
google-i18n-address==2.0.0
google-measurement-protocol==0.1.6
gunicorn==19.6.0
idna==2.1                 # via cryptography
jsonfield==1.0.3
lxml==3.7.0               # via premailer
Markdown==2.6.7
maxminddb-geolite2==2015.708
maxminddb==1.2.1
mock==2.0.0
pbr==1.10.0               # via mock
Pillow==3.4.2             # via django-versatileimagefield
premailer==3.0.1          # via django-emailit
prices==0.5.8
psycopg2==2.6.2
purl==1.3
py==1.4.31                # via pytest
pyasn1==0.1.9             # via cryptography
pycparser==2.17           # via cffi
PyJWT==1.4.2              # via django-payments
pytest-django==3.0.0
pytest==3.0.3
python-dateutil==2.5.3    # via fake-factory
pytz==2016.7              # via babel
redis==2.10.5             # via django-redis
requests==2.11.1
satchless==1.1.3
six==1.10.0               # via cryptography, fake-factory, mock, purl, python-dateutil
stripe==1.41.1            # via django-payments
suds-jurko==0.6           # via django-payments
unidecode==0.4.19
uwsgi==2.0.14
xmltodict==0.10.2         # via django-payments

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via cryptography
